started reorg of tree code
